requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
openai>=1.3.0
weaviate-client>=4.0.0
tqdm
pdfplumber
python-docx
numpy
markdownify
tabulate
simplejson
pytesseract>=0.3.10
pdf2image>=1.17.0
PyMuPDF>=1.23.9